+++ mdadash/backend/tests/test_server.py
@@ -604,6 +604,37 @@ async def test_widget_run_janin(_client, imd_server):
     await disconnect_from_simulation()
 
 
+async def test_widget_run_msd_serial(_client, imd_server):
+    uuid = await add_widget("MSD Analysis")
+    await connect_to_simulation(imd_server, step=1, batch_size=2)
+    inputs = [
+        ("selection", "resid 1"),
+        ("custom_title", ""),
+        ("show_particle_msds", True),
+        ("log_scale", False),
+    ]
+    await check_input_changes(uuid, inputs)
+    await resume_simulation(imd_server)
+    assert await sio_event_emitted(sio, "widgets:output", n=1)
+    await remove_widget(uuid)
+    await disconnect_from_simulation()
+
+
+async def test_widget_run_msd_parallel(_client, imd_server):
+    uuid = await add_widget("MSD Analysis")
+    await connect_to_simulation(imd_server, step=1, batch_size=2)
+    inputs = [
+        ("selection", "resid 1"),
+        ("show_particle_msds", True),
+        ("_run_mode", "parallel"),
+    ]
+    await check_input_changes(uuid, inputs)
+    await resume_simulation(imd_server)
+    assert await sio_event_emitted(sio, "widgets:output", n=1)
+    await remove_widget(uuid)
+    await disconnect_from_simulation()
+
+
 def test_state_load(tmp_path):
     # test with no state file
     sm = StateManager("")